Update paper.md
- added GM orcid and 2 affiliations
- added paragraphs on InterMine
- made small edits throughout the text to improve readability

The 2nd and 3rd InterMine paragraphs (starting: "Central to InterMine's interface..." and "Another feature of InterMine's interface") could be dropped or moved to Jose's ShEx paper though the second paragraph does introduce the concept of a template that is referred to later.

I need to add references (marked with **) and will do this next. This may explain why some parts are bold/italic when they should not be - apologies!